Add basic example of the libfive C++ API.
The primary reason for this example is because currently the test
suite for libfive compiles for WebAssembly but triggers an assertion.

This basic example allows demonstration that at some level compilation
for WebAssembly is possible even if the complete test suite doesn't
currently run.

It also serves as a complete example for the C++ API (including the
necessary includes etc) which the snippet on the example page isn't.